Remove use of non-existent metadata.thread.
The VME folks are using a label instead; Andrew and I discussed this
and don't think that's the right approach, although we don't have a
better solution for now, so just remove this and we will revisit it
later.

We can keep it in the VME version for now, but in general I'd strongly
prefer we prevent that from diverging from the "official" version.

Change-Id: I7c6eb05b569cfc345fd94127f736227e54fb5f48