Improve error handling for client authe errors.
There are other classes of "authentication error" aside from 'Invalid
Credentials', handle all of these and cause fluentd to retry (with backoff)
by re-raising the exception.

Also provide a better error message if we fail to fetch the service account
auth token, which can happen if no scopes are configured.