Parse container logs into a structured log if their body is json.
Treat them exactly the same as json logs from other sources if they are
json.

Avoid trying to parse every entry as json. If the first entry in a batch
isn't json, then don't try to parse the rest as json, since it's just a
waste of CPU. JSON parsing is unfortunately not cheap. We might consider
switching over to a faster parser (like https://github.com/ohler55/oj).