Detect JSON by checking each line, in all ingested logs (#77)
Instead of just checking the first line of each batch, check every line
for structured content in JSON format, for all ingested logs and not
just for logs coming from containers. This makes the logic easier to
understand and explain to users, and more uniform across the different
logs sources. The implementation now checks whether the start of the
line is valid JSON using simple string operations. Overall, assuming
that most batches of log lines are not JSON-formatted and have a
relatively low number of lines in the batch, this change should result
in less CPU compared to the old implementation which calls JSON.parse
and uses rescue to recover.