Request the logging scope from signet when fetching an access token.
The plugin already sets the logging scope to be requested for newly
acquired access token. However signet ignores this unless it obtains the
refresh token itself. google-fluentd usually uses a service account to
do so. If someone passes in an already created refresh token using the
GOOGLE_REFRESH_TOKEN environment variable, the access token request will
not request the logging scope. This fixes the latter.

More context about the change in signet (0.8+) can be found in
https://github.com/google/signet/pull/89