Fixed schema of OpenCensus metrics. (#353)
The previous version was using the same metric labels as the raw metrics in Prometheus. However, we need to write to the schema of the curated metrics. This change introduces the expected translation. This is only intended for legacy metrics; in the future we should make the raw metric equal to the curated metric.